Chore: Improved Javadocs and minor refactoring across classes
- Enhanced Javadoc in SparseSet for clarity.
- Removed dependency on EntityManager_ENTITY_LIMIT in SparseSet test.
- Added author tag with surname to FloatList.
- Corrected American spelling in ScalarStorage.
- Fixed issue in IntList: pop() was incorrectly returning a float instead of an int. No other classes or tests impacted (caught early).

+++ src/main/java/com/slinky/physics/util/IntList.java
@@ -429,7 +429,7 @@ public final class IntList {
      * @return the element previously at the last position in the list
      * @throws IndexOutOfBoundsException if the list is empty
      */
-    public float pop() {
+    public int pop() {
         return remove(size() - 1);
     }
